2-pull-e-otimizacao-de-prompts/src/utils.py
"""
Funções auxiliares para o projeto de otimização de prompts.
"""

import os
import yaml
import json
from typing import Dict, Any, Optional
from pathlib import Path
from dotenv import load_dotenv

load_dotenv()


def load_yaml(file_path: str) -> Optional[Dict[str, Any]]:
    """Carrega arquivo YAML."""
    try:
        with open(file_path, 'r', encoding='utf-8') as f:
            data = yaml.safe_load(f)
        return data
    except FileNotFoundError:
        print(f"Arquivo nao encontrado: {file_path}")
        return None
    except yaml.YAMLError as e:
        print(f"Erro ao parsear YAML: {e}")
        return None
    except Exception as e:
        print(f"Erro ao carregar arquivo: {e}")
        return None


def save_yaml(data: Dict[str, Any], file_path: str) -> bool:
    """Salva dados em arquivo YAML."""
    try:
        output_file = Path(file_path)
        output_file.parent.mkdir(parents=True, exist_ok=True)
        with open(output_file, 'w', encoding='utf-8') as f:
            yaml.dump(data, f, allow_unicode=True, sort_keys=False, indent=2)
        return True
    except Exception as e:
        print(f"Erro ao salvar arquivo: {e}")
        return False


def check_env_vars(required_vars: list) -> bool:
    """Verifica se variáveis de ambiente obrigatórias estão configuradas."""
    missing_vars = [var for var in required_vars if not os.getenv(var)]
    if missing_vars:
        print("Variaveis de ambiente faltando:")
        for var in missing_vars:
            print(f"   - {var}")
        print("\nConfigure-as no arquivo .env antes de continuar.")
        return False
    return True


def format_score(score: float, threshold: float = 0.9) -> str:
    """Formata score com indicador visual."""
    symbol = "✓" if score >= threshold else "✗"
    return f"{score:.2f} {symbol}"


def print_section_header(title: str, char: str = "=", width: int = 50):
    """Imprime cabeçalho de seção formatado."""
    print("\n" + char * width)
    print(title)
    print(char * width + "\n")


def validate_prompt_structure(prompt_data: Dict[str, Any]) -> tuple:
    """Valida estrutura básica de um prompt."""
    errors = []
    required_fields = ['description', 'system_prompt', 'version']
    for field in required_fields:
        if field not in prompt_data:
            errors.append(f"Campo obrigatorio faltando: {field}")

    system_prompt = prompt_data.get('system_prompt', '').strip()
    if not system_prompt:
        errors.append("system_prompt esta vazio")
    if 'TODO' in system_prompt:
        errors.append("system_prompt ainda contem TODOs")

    techniques = prompt_data.get('techniques_applied', [])
    if len(techniques) < 2:
        errors.append(f"Minimo de 2 tecnicas requeridas, encontradas: {len(techniques)}")

    return (len(errors) == 0, errors)


def extract_json_from_response(response_text: str) -> Optional[Dict[str, Any]]:
    """Extrai JSON de uma resposta de LLM."""
    try:
        return json.loads(response_text)
    except json.JSONDecodeError:
        start = response_text.find('{')
        end = response_text.rfind('}') + 1
        if start != -1 and end > start:
            try:
                return json.loads(response_text[start:end])
            except json.JSONDecodeError:
                pass
    return None


def get_llm(model: Optional[str] = None, temperature: float = 0.0):
    """Retorna instância de LLM configurada baseada no provider."""
    provider = os.getenv('LLM_PROVIDER', 'openai').lower()
    model_name = model or os.getenv('LLM_MODEL', 'gpt-4o-mini')

    if provider == 'openai':
        from langchain_openai import ChatOpenAI
        api_key = os.getenv('OPENAI_API_KEY')
        if not api_key:
            raise ValueError("OPENAI_API_KEY nao configurada no .env")
        return ChatOpenAI(model=model_name, temperature=temperature, api_key=api_key)

    elif provider in ('google', 'gemini'):
        from langchain_google_genai import ChatGoogleGenerativeAI
        api_key = os.getenv('GOOGLE_API_KEY')
        if not api_key:
            raise ValueError("GOOGLE_API_KEY nao configurada no .env")
        return ChatGoogleGenerativeAI(
            model=model_name, temperature=temperature, google_api_key=api_key
        )
    else:
        raise ValueError(f"Provider '{provider}' nao suportado. Use 'openai' ou 'google'.")


def get_eval_llm(temperature: float = 0.0):
    """Retorna LLM configurado para avaliação (usa EVAL_MODEL)."""
    eval_model = os.getenv('EVAL_MODEL', 'gpt-4o')
    return get_llm(model=eval_model, temperature=temperature)
