2-pull-e-otimizacao-de-prompts/requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-openai
langsmith
python-dotenv
pyyaml
pytest
